RepBot-v1.02/requirements.txt
discord.py>=2.0.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
pytz>=2023.3